+++ src/test/java/btrplace/plan/event/ShutdownVMTest.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package btrplace.plan.action;
+package btrplace.plan.event;
 
 import btrplace.model.DefaultMapping;
 import btrplace.model.DefaultModel;
@@ -28,22 +28,19 @@ import org.testng.annotations.Test;
 import java.util.UUID;
 
 /**
- * Unit tests for {@link MigrateVM}.
+ * Unit tests for {@link ShutdownVM}.
  *
  * @author Fabien Hermenier
  */
-public class MigrateVMTest {
+public class ShutdownVMTest {
 
     @Test
     public void testInstantiate() {
         UUID vm = UUID.randomUUID();
-        UUID n1 = UUID.randomUUID();
-        UUID n2 = UUID.randomUUID();
-
-        MigrateVM a = new MigrateVM(vm, n1, n2, 3, 5);
+        UUID n = UUID.randomUUID();
+        ShutdownVM a = new ShutdownVM(vm, n, 3, 5);
         Assert.assertEquals(vm, a.getVM());
-        Assert.assertEquals(n1, a.getSourceNode());
-        Assert.assertEquals(n2, a.getDestinationNode());
+        Assert.assertEquals(n, a.getNode());
         Assert.assertEquals(3, a.getStart());
         Assert.assertEquals(5, a.getEnd());
         Assert.assertFalse(a.toString().contains("null"));
@@ -53,55 +50,39 @@ public class MigrateVMTest {
     @Test(dependsOnMethods = {"testInstantiate"})
     public void testApply() {
         Mapping map = new DefaultMapping();
-        UUID vm = UUID.randomUUID();
-        UUID n1 = UUID.randomUUID();
-        UUID n2 = UUID.randomUUID();
-
-        map.addOnlineNode(n1);
-        map.addOnlineNode(n2);
-        map.addRunningVM(vm, n1);
-
         Model m = new DefaultModel(map);
-
-        MigrateVM a = new MigrateVM(vm, n1, n2, 3, 5);
+        UUID vm = UUID.randomUUID();
+        UUID n = UUID.randomUUID();
+        ShutdownVM a = new ShutdownVM(vm, n, 3, 5);
+        map.addOnlineNode(n);
+        map.addRunningVM(vm, n);
         Assert.assertTrue(a.apply(m));
-        Assert.assertEquals(map.getVMLocation(vm), n2);
-        Assert.assertTrue(map.getRunningVMs().contains(vm));
+        Assert.assertTrue(map.getReadyVMs().contains(vm));
 
         Assert.assertFalse(a.apply(m));
-        Assert.assertEquals(map.getVMLocation(vm), n2);
 
-        Assert.assertFalse(new MigrateVM(vm, n2, n2, 3, 5).apply(m));
+        map.addSleepingVM(vm, n);
+        Assert.assertFalse(a.apply(m));
 
-        map.addSleepingVM(vm, n2);
-        Assert.assertFalse(new MigrateVM(vm, n2, n1, 3, 5).apply(m));
+        map.removeVM(vm);
+        Assert.assertFalse(a.apply(m));
 
         map.addReadyVM(vm);
-        Assert.assertFalse(new MigrateVM(vm, n2, n1, 3, 5).apply(m));
-
-        map.addOfflineNode(n1);
-        Assert.assertFalse(new MigrateVM(vm, n2, n1, 3, 5).apply(m));
-
-        map.removeNode(n1);
-        Assert.assertFalse(new MigrateVM(vm, n2, n1, 3, 5).apply(m));
+        map.addOfflineNode(n);
+        Assert.assertFalse(a.apply(m));
     }
 
     @Test(dependsOnMethods = {"testInstantiate"})
     public void testEquals() {
+        UUID n = UUID.randomUUID();
         UUID vm = UUID.randomUUID();
-        UUID n1 = UUID.randomUUID();
-        UUID n2 = UUID.randomUUID();
-
-        MigrateVM a = new MigrateVM(vm, n1, n2, 3, 5);
-        MigrateVM b = new MigrateVM(vm, n1, n2, 3, 5);
+        ShutdownVM a = new ShutdownVM(vm, n, 3, 5);
+        ShutdownVM b = new ShutdownVM(vm, n, 3, 5);
         Assert.assertEquals(a, b);
         Assert.assertEquals(a.hashCode(), b.hashCode());
-
-        Assert.assertNotSame(a, new MigrateVM(vm, n1, n2, 4, 5));
-        Assert.assertNotSame(a, new MigrateVM(vm, n1, n2, 3, 4));
-        Assert.assertNotSame(a, new MigrateVM(UUID.randomUUID(), n1, n2, 3, 5));
-        Assert.assertNotSame(a, new MigrateVM(vm, UUID.randomUUID(), n2, 3, 5));
-        Assert.assertNotSame(a, new MigrateVM(vm, n1, UUID.randomUUID(), 3, 5));
-
+        Assert.assertNotSame(a, new ShutdownVM(vm, n, 4, 5));
+        Assert.assertNotSame(a, new ShutdownVM(vm, n, 3, 4));
+        Assert.assertNotSame(a, new ShutdownVM(vm, UUID.randomUUID(), 3, 5));
+        Assert.assertNotSame(a, new ShutdownVM(UUID.randomUUID(), n, 4, 5));
     }
 }